requirements.txt
filterpy==1.4.5
loguru==0.6.0
matplotlib==3.3.4
numpy==1.19.5
onnx==1.8.1
onnxruntime==1.11.1
onnxruntime_gpu==1.8.0
opencv_python==4.5.5.64
pycocotools==2.0.4
pyserial==3.5
tabulate==0.8.9
thop==0.0.31.post2005241907
torch==1.8.0
torchvision==0.12.0
tqdm==4.64.0
